Feature: Move open at startup to advanced settings
