+++ src/Files.App/ViewModels/Settings/GeneralViewModel.cs
@@ -5,7 +5,6 @@ using Files.Shared.Services.DateTimeFormatter;
 using Microsoft.Extensions.Logging;
 using System.Collections.Specialized;
 using System.Globalization;
-using Windows.ApplicationModel;
 using Windows.Globalization;
 using Windows.Storage;
 using Windows.Storage.Pickers;
@@ -23,7 +22,6 @@ namespace Files.App.ViewModels.Settings
 
 		private ReadOnlyCollection<IMenuFlyoutItemViewModel> addFlyoutItemsSource;
 
-		public AsyncRelayCommand OpenFilesAtStartupCommand { get; }
 		public AsyncRelayCommand ChangePageCommand { get; }
 		public RelayCommand RemovePageCommand { get; }
 		public RelayCommand<string> AddPageCommand { get; }
@@ -93,7 +91,6 @@ namespace Files.App.ViewModels.Settings
 
 		public GeneralViewModel()
 		{
-			OpenFilesAtStartupCommand = new AsyncRelayCommand(OpenFilesAtStartup);
 			ChangePageCommand = new AsyncRelayCommand(ChangePage);
 			RemovePageCommand = new RelayCommand(RemovePage);
 			AddPageCommand = new RelayCommand<string>(async (path) => await AddPage(path));
@@ -389,79 +386,6 @@ namespace Files.App.ViewModels.Settings
 					OnPropertyChanged();
 				}
 			}
-		}
-
-		private bool openInLogin;
-		public bool OpenInLogin
-		{
-			get => openInLogin;
-			set => SetProperty(ref openInLogin, value);
-		}
-
-		private bool canOpenInLogin;
-		public bool CanOpenInLogin
-		{
-			get => canOpenInLogin;
-			set => SetProperty(ref canOpenInLogin, value);
-		}
-
-		public async Task OpenFilesAtStartup()
-		{
-			var stateMode = await ReadState();
-
-			bool state = stateMode switch
-			{
-				StartupTaskState.Enabled => true,
-				StartupTaskState.EnabledByPolicy => true,
-				StartupTaskState.DisabledByPolicy => false,
-				StartupTaskState.DisabledByUser => false,
-				_ => false,
-			};
-
-			if (state != OpenInLogin)
-			{
-				StartupTask startupTask = await StartupTask.GetAsync("3AA55462-A5FA-4933-88C4-712D0B6CDEBB");
-				if (OpenInLogin)
-					await startupTask.RequestEnableAsync();
-				else
-					startupTask.Disable();
-				await DetectOpenFilesAtStartup();
-			}
-		}
-
-		public async Task DetectOpenFilesAtStartup()
-		{
-			var stateMode = await ReadState();
-
-			switch (stateMode)
-			{
-				case StartupTaskState.Disabled:
-					CanOpenInLogin = true;
-					OpenInLogin = false;
-					break;
-				case StartupTaskState.Enabled:
-					CanOpenInLogin = true;
-					OpenInLogin = true;
-					break;
-				case StartupTaskState.DisabledByPolicy:
-					CanOpenInLogin = false;
-					OpenInLogin = false;
-					break;
-				case StartupTaskState.DisabledByUser:
-					CanOpenInLogin = false;
-					OpenInLogin = false;
-					break;
-				case StartupTaskState.EnabledByPolicy:
-					CanOpenInLogin = false;
-					OpenInLogin = true;
-					break;
-			}
-		}
-
-		public async Task<StartupTaskState> ReadState()
-		{
-			var state = await StartupTask.GetAsync("3AA55462-A5FA-4933-88C4-712D0B6CDEBB");
-			return state.State;
 		}
 
 		public bool SearchUnindexedItems
